Create README.md
Using a total of 11 functions in python, Roman numerals are entered by the user and tested for validity within the program. Then, the number eventually is converted to an arabic symbol(numeric number) and printed out to the user. The program works both ways so the user can also enter a numeric number(arabic symbol) and get returned as a roman numeral.